Move TimeStampedModel to base app.
This commit also updated the corresponding import statements.
Earlier this model was present in accounts app.
