+++ apps/accounts/models.py
@@ -3,18 +3,7 @@ from __future__ import unicode_literals
 from django.db import models
 from django.contrib.auth.models import User
 
-
-class TimeStampedModel(models.Model):
-    """
-    An abstract base class model that provides self-managed `created_at` and
-    `modified_at` fields.
-    """
-    created_at = models.DateTimeField(auto_now_add=True)
-    modified_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        abstract = True
-        app_label = 'accounts'
+from base.models import (TimeStampedModel,)
 
 
 class UserStatus(TimeStampedModel):
