+++ apps/base/models.py
@@ -2,4 +2,15 @@ from __future__ import unicode_literals
 
 from django.db import models
 
-# Create your models here.
+
+class TimeStampedModel(models.Model):
+    """
+    An abstract base class model that provides self-managed `created_at` and
+    `modified_at` fields.
+    """
+    created_at = models.DateTimeField(auto_now_add=True)
+    modified_at = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        abstract = True
+        app_label = 'base'
